Add confirmation options, update-books options and adapt old jnc.py
